refactor(app): extract Command into app/command.rs
Moved verbatim and re-exported. Part of #270.
